Refactor blog post page and improve markdown support
Extracted blog post content into a new BlogPostContent component for better separation of concerns. Enhanced MarkdownPreview to support images and improved code block styling. Updated MarkdownEditor to add an image button. Improved hooks for fetching blog data by using useCallback and destructuring options. Added custom prose styles for blog content in globals.css.
