+++ app/globals.css
@@ -309,3 +309,21 @@
 .mask-rect:nth-child(4) {
   --animation-delay: 3s;
 }
+
+/* Custom prose styles for blog content */
+.prose a {
+  color: inherit !important;
+  text-decoration: underline;
+  text-decoration-color: currentColor;
+}
+
+.prose a:hover {
+  text-decoration: none;
+}
+
+.prose img {
+  border-radius: 0.5rem;
+  margin: 1rem 0;
+  max-width: 100%;
+  height: auto;
+}
